Struct fixedvec::Iter

pub struct Iter<'a, T> where T: 'a {
    // some fields omitted
}

Immutable slice iterator

Examples

Basic usage:

// First, we declare a type which has `iter` method to get the `Iter` struct (&[usize here]):
let slice = &[1, 2, 3];

// Then, we iterate over it:
for element in slice.iter() {
    println!("{}", element);
}

Methods

impl<'a, T> Iter<'a, T>

fn as_slice(&self) -> &'a [T]

View the underlying data as a subslice of the original data.

This has the same lifetime as the original slice, and so the iterator can continue to be used while this exists.

Examples

Basic usage:

// First, we declare a type which has the `iter` method to get the `Iter`
// struct (&[usize here]):
let slice = &[1, 2, 3];

// Then, we get the iterator:
let mut iter = slice.iter();
// So if we print what `as_slice` method returns here, we have "[1, 2, 3]":
println!("{:?}", iter.as_slice());

// Next, we move to the second element of the slice:
iter.next();
// Now `as_slice` returns "[2, 3]":
println!("{:?}", iter.as_slice());